Dunnjacob (#5)
* [Update] adds axios and map recipe logic

* [Testing] Updates testing

* feature: Componetizes Button and Carousel

* fixes: merge conflicts
