+++ client/src/components/Button.js
@@ -1,0 +1,17 @@
+import React from "react";
+
+function Button(props) {
+  return (
+    <button
+      value={props.value}
+      type={props.type}
+      className={props.className}
+      onClick={props.onClick}
+      //size={props?.size}
+    >
+      {props.title}
+    </button>
+  );
+}
+
+export default Button;
